feat(cpp): split header file (#113)
