+++ android/src/main/jni.cpp
@@ -11,7 +11,7 @@
 #include "llama.h"
 #include "llama-impl.h"
 #include "ggml.h"
-#include "rn-llama.hpp"
+#include "rn-llama.h"
 #include "jni-utils.h"
 
 #define UNUSED(x) (void)(x)
